Limpieza en componentes de gestion de turnos y dashboard
