+++ dental-landing/src/Render/components/UI/index.ts
@@ -1,0 +1,3 @@
+export * from './Button/Button';
+export * from './GoBack/GoBack';
+export * from './Header/Header';
