+++ dental-landing/src/Render/pages/Reports/Statistics/AppointmentsStatistics/AppointmentsByConfirm/AppointmentsByConfirm.tsx
