+++ dental-landing/src/Render/pages/Reports/Statistics/PatientsStatistics/PatientsByGenderChart/PatientsByGenderChart.tsx
