+++ dental-landing/src/Render/pages/Reports/Statistics/PatientsStatistics/PatientsByTreatmentChart/PatientsByTreatmentChart.tsx
