+++ dental-landing/src/Render/pages/Reports/Statistics/index.ts
@@ -1,0 +1,7 @@
+export * from './AppointmentsStatistics/AppointmentsByConfirm/AppointmentsByConfirm';
+export * from './AppointmentsStatistics/AppointmentsByMonthChart/AppointmentsByMonthChart';
+export * from './AppointmentsStatistics/AppointmentsStatistics';
+export * from './PatientsStatistics/PatientsByAgeChart/PatientsByAgeChart';
+export * from './PatientsStatistics/PatientsByGenderChart/PatientsByGenderChart';
+export * from './PatientsStatistics/PatientsByTreatmentChart/PatientsByTreatmentChart';
+export * from './PatientsStatistics/PatientsStatistics';
