+++ dental-landing/src/Render/pages/UserManagement/EditUserInfo.tsx
@@ -1,11 +1,12 @@
-import { Link, useParams } from 'react-router-dom';
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { useParams } from 'react-router-dom';
 import Navbar from '../../components/Platform/Navbar';
-import { ChevronLeftIcon } from '@heroicons/react/20/solid';
 import { useEffect, useState } from 'react';
 import { useForm } from 'react-hook-form';
 import axios, { AxiosError } from 'axios';
 import Swal from 'sweetalert2';
 import { token } from '../../../localStorage/token';
+import { GoBack } from '../../components/UI';
 
 export const EditUserInfo = () => {
   const { id } = useParams();
@@ -74,17 +75,7 @@ export const EditUserInfo = () => {
     <>
       <Navbar />
       <section className="mx-auto lg:ms-[200px] max-w-[1594px] mt-[150px]">
-        <div className="flex items-center mb-6">
-          <Link to="/users-management/users-list" className="me-16">
-            <button className="flex items-center bg-transparent poppins-medium">
-              <ChevronLeftIcon
-                className="h-5 w-5 flex-none text-black"
-                aria-hidden="true"
-              />
-              Atrás
-            </button>
-          </Link>
-        </div>
+        <GoBack path='users-management/users-list' titleGoBack='Atrás'/>
         <main className="w-[90%] max-w-[1594px] h-[740px] rounded-[35px] bg-lightgray border border-[#424242] py-[30px] px-[78px] mx-auto">
           <form onSubmit={handleSubmit(onSubmit)}>
             <div className="flex flex-col justify-center items-center poppins-light text-[16px] space-y-4">
