--- dental-landing/src/types/dtos/patient/create-patient.type.ts
@@ -1,12 +1,0 @@
-// export interface Patient {
-//   id: number;
-//   name: string;
-//   surname: string;
-//   dni: number;
-//   gender: 'Male' | 'Female';
-//   adress: string;
-//   pEmail: string;
-//   phone: number;
-//   appointments: [];
-//   medicalHistories: [];
-// }
