+++ dental-landing/src/types/props/goBack.props.ts
@@ -1,0 +1,4 @@
+export interface GoBackProps {
+    path: string
+    titleGoBack: string
+}
